refactor(studio-integration): Move Default Mock Content Provider
Now available in MockContentPlugin.
